+++ src/components/Loaders/Friends.jsx
